requirements.txt
beautifulsoup4
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
rich